refactor unit tests into dir
