+++ crates/dex-orderbook/src/tests/mod.rs
@@ -1,0 +1,5 @@
+/// Unit tests
+pub mod basic;
+pub mod fuzz;
+pub mod swap_math;
+pub mod test_utils;
